Add return statements to subroutines
I was under the impression that "return" was a bashism, but I was wrong.
I was getting failures until I removed them, but for some other reason.
Sorry darrob.

I'm adding the return statements back. They work work bash, dash, and even
the completely anal "posh", and it seems that without out them the exit
code is nearly always non-zero with set -e.