Make sure NEWS.html exists
Make sure NEWS.html exists on every run of the script. If it doesn't
exist it creates a simple placeholder informing the user about
grid-updates. This is necessary because they might be using a patched
Tahoe-LAFS version that expects to display NEWS.html in an Iframe on its
WebUI front page.