Update default NEWSURI
The default location in older versions will still work.

Reason for change: in the old location we used `tahoe backup` to update
the NEWS file and keep all previous versions in an archive directory.
Since our switch to the binary NEWS.tgz format this doesn't seem like a
good idea anymore, because the archive can't be easily browsed.

If we still really wanted an archive, we should only keep the plain-text
versions in it.

Not having an archive at all will (at least theoretically) speed up the
repair process.