filter backed-up introducers before comparing
The "if diff" line is changed to "if not diff" because diff exits 1 if
there are changes and exits with status 0 if there are not.

Also, we want to compare the old introducer list with the new one. The
way it was before, we'd compare the backup with the introducer list; the
introducer list should be the same as the backup--we just backed it up!