Fix crash on repair of LIT files
Literal files cannot be repaired and therefor return very different JSON
results.

I hope this deals with them suffiently but it's possible that just
checking for an empty 'storage-index' in not enough. In that case
is_literal_file() could be extended to also parse the 'cap' which begins
with URI:LIT:.