Correct HTML manpage
When in manpage form, the AUTHOR section is *below* the BUGS section.
Once compiled to HTML, the author section moves to the top of the file.
Therefore, we'll make a link that's anchored to the header id.

Also, let's turn off the obsfucation of email addresses; it's off for
README & INSTALL.